# contract/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: C15Tour API
  version: 0.0.1
  description: API minimale pour initialisation.

paths:
  /ping:
    get:
      summary: Health Check
      operationId: ping
      tags:
        - System
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /tours:
    post:
      summary: Create a new tour
      operationId: createTour
      tags:
        - Tours
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TourCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TourResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: Get all tours
      operationId: getAllTours
      tags:
        - Tours
      responses:
        '200':
          description: List of tours
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TourResponse'
  /tours/{id}:
    get:
      summary: Get a tour by ID
      operationId: getTourById
      tags:
        - Tours
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Tour found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TourResponse'
        '404':
          description: Tour not found

    put:
      summary: Update a tour by ID
      operationId: updateTour
      tags:
        - Tours
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TourCreateRequest'
      responses:
        '200':
          description: Tour updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TourResponse'
        '404':
          description: Tour not found

    delete:
      summary: Delete a tour by ID
      operationId: deleteTour
      tags:
        - Tours
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Tour deleted
        '404':
          description: Tour not found
  /tours/share/{code}:
    get:
      summary: Get a tour by share code
      operationId: getTourByShareCode
      tags:
        - Tours
      parameters:
        - name: code
          in: path
          required: true
          description: "6character unique share code"
          schema:
            type: string
      responses:
        '200':
          description: Tour found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TourResponse'
        '404':
          description: Tour not found

  /tours/share/{code}/join:
    post:
      summary: Join a tour as organiser or participant
      operationId: joinTour
      tags:
        - Mobile
      parameters:
        - name: code
          in: path
          required: true
          description: "6-character share code (participant) or organiser code"
          schema:
            type: string
      responses:
        '200':
          description: Joined successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinResponse'
        '404':
          description: Invalid code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Organiser already joined
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /tours/share/{code}/organiser-position:
    put:
      summary: Update the organiser's current position
      operationId: updateOrganiserPosition
      tags:
        - Mobile
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
        - name: X-Session-Token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganiserPositionRequest'
      responses:
        '204':
          description: Position updated
        '401':
          description: Invalid or missing session token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Tour not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /tours/{id}/route-to-start:
    post:
      summary: Get route from current location to the start of the tour
      operationId: getRouteToStart
      tags:
        - TourCalculation
      parameters:
        - name: id
          in: path
          required: true
          description: "ID du tour"
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPositionRequest'
      responses:
        '200':
          description: Route calculated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteToStartResponse'
        '404':
          description: Tour not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Coordinates:
      type: object
      required:
        - latitude
        - longitude
      properties:
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
          example: 47.218371
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
          example: -1.553621

    Message:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          example: "Pong"

    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: "Invalid input data"

    Waypoints:
      type: object
      required:
        - name
        - coordinates
      properties:
        name:
          type: string
          example: "Etape 1"
        coordinates:
          $ref: "#/components/schemas/Coordinates"
        estimatedArrival:
          type: string
          format: date-time
          description: Computed estimated arrival time at this waypoint (only set when departureTime is provided)
          example: "2024-06-01T14:30:00Z"

    SegmentRequest:
      type: object
      required:
        - name
        - waypoints
      properties:
        name:
          type: string
          example: "Segment 1"
        waypoints:
          type: array
          minItems: 2
          items:
            $ref: "#/components/schemas/Waypoints"
        breakDuration:
          type: integer
          description: How long to stay at the end of this segment in seconds (default 0)
          example: 3600

    SegmentResponse:
      type: object
      required:
        - name
        - distance
        - duration
        - geometry
        - waypoints
      properties:
        name:
          type: string
          example: "Segment 1"
        distance:
          type: integer
          format: int64
          description: Total distance in meters
          example: 256340
        duration:
          type: integer
          format: int64
          description: Total duration in seconds
          example: 4248
        geometry:
          type: string
          description: Geometry of the OSRM Response on type JSON
        steps:
          type: string
          description: List of navigation steps and maneuvers (JSON string)
        waypoints:
          type: array
          description: List of the points which compose the segment
          items:
            $ref: "#/components/schemas/Waypoints"
        breakDuration:
          type: integer
          description: How long to stay at the end of this segment in seconds
          example: 3600
        estimatedDeparture:
          type: string
          format: date-time
          description: Computed estimated departure time from the end of this segment (arrival at last waypoint + break duration)
          example: "2024-06-01T16:30:00Z"

    TourCreateRequest:
      type: object
      required:
        - name
        - segments
      properties:
        name:
          type: string
          example: "Roadtrip Nantes"
        segments:
          type: array
          items:
            $ref: '#/components/schemas/SegmentRequest'
        departureTime:
          type: string
          format: date-time
          description: When the tour starts (optional, required for ETA computation)
          example: "2024-06-01T08:00:00Z"

    TourResponse:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Roadtrip Nantes"
        shareCode:
          type: string
          example: "C15ROC"
        organiserCode:
          type: string
          example: "ORG123"
          description: Only returned to admin — never expose to participants
        totalDistance:
          type: integer
          description: Total distance of the tour in meters
          example: 50
        totalDuration:
          type: integer
          description: Total duration of the tour in seconds
          example: 3852
        createdAt:
          type: string
          format: date-time
          example: "2024-06-01T12:00:00Z"
          description: Tour creation date and time
        departureTime:
          type: string
          format: date-time
          example: "2024-06-01T08:00:00Z"
          description: When the tour starts (optional)
        segments:
          type: array
          items:
            $ref: '#/components/schemas/SegmentResponse'

    JoinResponse:
      type: object
      required:
        - role
      properties:
        role:
          type: string
          enum: [ORGANISER, PARTICIPANT]
        sessionToken:
          type: string
          description: Only present when role is ORGANISER

    OrganiserPositionRequest:
      type: object
      required:
        - latitude
        - longitude
      properties:
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180

    UserPositionRequest:
      type: object
      required:
        - coordinates
      properties:
        coordinates:
          $ref: '#/components/schemas/Coordinates'

    RouteToStartResponse:
      type: object
      required:
        - distance
        - duration
        - geometry
      properties:
        distance:
          type: integer
          format: int64
          description: Distance en mètres depuis la position actuelle jusqu'au début du tour
          example: 5000
        duration:
          type: integer
          format: int64
          description: Durée en secondes depuis la position actuelle jusqu'au début du tour
          example: 600
        geometry:
          type: string
          description: Geometry de la route au format JSON (OSRM)
        steps:
          type: string
          description: List des étapes de navigation (JSON string)
        startPoint:
          $ref: '#/components/schemas/Waypoints'